+++ CHANGES.md
@@ -2,9 +2,12 @@
 
 ## 1.0.0: 19 Feb 2019
 - RENAME main class MWClient --> MWSite
-- REFACT: SPLIT classes across `HttpPipeline`, `HttpCall`, `HttpSession`(handlers) & `MwSite`,
-  so client operations can inspect and modify request & response with callbaks (handlers)
-  (ala [apache axis](http://axis.apache.org/axis2/java/core/docs/userguide.html#handlessoap)).
+- REFACT: 
+  - SPLIT classes across `HttpPipeline`, `HttpCall`, `HttpSession`(handlers) & `MwSite`,
+    so client operations can inspect and modify request & response with callbaks (handlers)
+    (ala [apache axis](http://axis.apache.org/axis2/java/core/docs/userguide.html#handlessoap)).
+- UNIFY UNIFY HttpCall() + MWSite/apiCall() arg-parsing (uri, uriArgs, method, headers, body, hoptions)
+  - `MWSite.apiCall()` accepts kv-pairs (see `HttpCall.inputParser` & `HttpCall` constructor).
 - Refact DatumError --> MWError:
   - rename property Datum --> HttpCall.
   - store the hole call instead of just the response/request.